requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
better-profanity==0.7.0
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
discord.py==1.7.2
Flask==2.0.1
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
jws==0.1.3
MarkupSafe==2.0.1
multidict==5.1.0
oauth2client==3.0.0
protobuf==3.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dotenv==0.17.1
python-jwt==2.0.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.7.2
six==1.16.0
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==2.0.1
yarl==1.6.3
